enhance: resize high priority wqthreadpool dynamically(#40838) (#41549)
related: #40838

+++ internal/core/src/common/Common.cpp
@@ -20,11 +20,11 @@
 namespace milvus {
 
 int64_t FILE_SLICE_SIZE = DEFAULT_INDEX_FILE_SLICE_SIZE;
-int64_t HIGH_PRIORITY_THREAD_CORE_COEFFICIENT =
+float HIGH_PRIORITY_THREAD_CORE_COEFFICIENT =
     DEFAULT_HIGH_PRIORITY_THREAD_CORE_COEFFICIENT;
-int64_t MIDDLE_PRIORITY_THREAD_CORE_COEFFICIENT =
+float MIDDLE_PRIORITY_THREAD_CORE_COEFFICIENT =
     DEFAULT_MIDDLE_PRIORITY_THREAD_CORE_COEFFICIENT;
-int64_t LOW_PRIORITY_THREAD_CORE_COEFFICIENT =
+float LOW_PRIORITY_THREAD_CORE_COEFFICIENT =
     DEFAULT_LOW_PRIORITY_THREAD_CORE_COEFFICIENT;
 int CPU_NUM = DEFAULT_CPU_NUM;
 int64_t EXEC_EVAL_EXPR_BATCH_SIZE = DEFAULT_EXEC_EVAL_EXPR_BATCH_SIZE;
@@ -40,21 +40,21 @@ SetIndexSliceSize(const int64_t size) {
 }
 
 void
-SetHighPriorityThreadCoreCoefficient(const int64_t coefficient) {
+SetHighPriorityThreadCoreCoefficient(const float coefficient) {
     HIGH_PRIORITY_THREAD_CORE_COEFFICIENT = coefficient;
     LOG_INFO("set high priority thread pool core coefficient: {}",
              HIGH_PRIORITY_THREAD_CORE_COEFFICIENT);
 }
 
 void
-SetMiddlePriorityThreadCoreCoefficient(const int64_t coefficient) {
+SetMiddlePriorityThreadCoreCoefficient(const float coefficient) {
     MIDDLE_PRIORITY_THREAD_CORE_COEFFICIENT = coefficient;
     LOG_INFO("set middle priority thread pool core coefficient: {}",
              MIDDLE_PRIORITY_THREAD_CORE_COEFFICIENT);
 }
 
 void
-SetLowPriorityThreadCoreCoefficient(const int64_t coefficient) {
+SetLowPriorityThreadCoreCoefficient(const float coefficient) {
     LOW_PRIORITY_THREAD_CORE_COEFFICIENT = coefficient;
     LOG_INFO("set low priority thread pool core coefficient: {}",
              LOW_PRIORITY_THREAD_CORE_COEFFICIENT);
